Fixed STM platforms IRQ handlers when using a C++ compiler
